Refactor ML model data classes
