+++ backend/server/ml_models/all_model/all_model.py
@@ -1,6 +1,6 @@
 """Class for model trained on all AFL data and its associated data class"""
 
-from typing import List, Tuple, Union, Callable, Sequence, Optional
+from typing import List, Callable, Sequence, Optional
 from datetime import datetime
 from functools import reduce
 import pandas as pd
@@ -12,8 +12,8 @@ from xgboost import XGBRegressor
 from server.ml_models.betting_model import BettingModelData
 from server.ml_models.match_model import MatchModelData
 from server.ml_models.player_model import PlayerModelData
-from server.ml_models.ml_model import MLModel
-from server.types import YearsType
+from server.ml_models.ml_model import MLModel, MLModelData
+from server.types import YearPair
 
 
 START_DATE = "1965-01-01"
@@ -40,33 +40,18 @@ class AllModel(MLModel):
         super().__init__(estimators=estimators, name=name, module_name=module_name)
 
 
-class AllModelData:
-    """Load and clean data for the XGBRegressor pipeline.
-
-    Args:
-        train_years (tuple[integer or None]): Minimum and maximum (inclusive) years
-            for the training data.
-        test_years (tuple[ingeter or None]): Minimum and maximum (inclusive) years
-            for the test data.
-
-    Attributes:
-        data (pandas.DataFrame): Cleaned, unfiltered data frame.
-        train_years (tuple[integer or None]): Minimum and maximum (inclusive) years
-            for the training data.
-        test_years (tuple[ingeter or None]): Minimum and maximum (inclusive) years
-            for the test data.
-    """
+class AllModelData(MLModelData):
+    """Load and clean data from all data sources"""
 
     def __init__(
         self,
         data_readers: List[Callable] = DATA_READERS,
-        train_years: YearsType = (None, 2015),
-        test_years: YearsType = (2016, 2016),
+        train_years: YearPair = (None, 2015),
+        test_years: YearPair = (2016, 2016),
         start_date=None,
         end_date="2016-12-31",
     ) -> None:
-        self._train_years = train_years
-        self._test_years = test_years
+        super().__init__(train_years=train_years, test_years=test_years)
 
         data_frame = reduce(self.__concat_data_frames, data_readers)
         data_frame_dtypes = data_frame.dtypes
@@ -79,7 +64,7 @@ class AllModelData:
             datetime.strptime(end_date, "%Y-%m-%d").year if start_date else np.Inf
         )
 
-        self.data = (
+        self._data = (
             data_frame[
                 (data_frame["year"] >= start_year) & (data_frame["year"] <= end_year)
             ]
@@ -87,95 +72,9 @@ class AllModelData:
             .dropna()
         )
 
-    def train_data(self) -> Tuple[pd.DataFrame, pd.DataFrame]:
-        """Filter data by year to produce training data.
-
-        Returns:
-            Tuple[pandas.DataFrame]: Training features and labels.
-        """
-
-        data_train = self.data[
-            (self.data["year"] >= self.__train_min())
-            & (self.data["year"] <= self.__train_max())
-        ]
-
-        X_train = self.__X(data_train)
-        y_train = self.__y(data_train)
-
-        return X_train, y_train
-
-    def test_data(self) -> Tuple[pd.DataFrame, pd.DataFrame]:
-        """Filter data by year to produce test data.
-
-        Returns:
-            Tuple[pandas.DataFrame]: Test features and labels.
-        """
-
-        data_test = self.data[
-            (self.data["year"] >= self.__test_min())
-            & (self.data["year"] <= self.__test_max())
-        ]
-        X_test = self.__X(data_test)
-        y_test = self.__y(data_test)
-
-        return X_test, y_test
-
     @property
-    def train_years(self) -> YearsType:
-        """Range of years for slicing training data"""
-
-        return self._train_years
-
-    @train_years.setter
-    def train_years(self, years: YearsType) -> None:
-        self._train_years = years
-
-    @property
-    def test_years(self) -> YearsType:
-        """Range of years for slicing test data"""
-
-        return self._test_years
-
-    @test_years.setter
-    def test_years(self, years: YearsType) -> None:
-        self._test_years = years
-
-    def __train_min(self) -> Union[int, float]:
-        return self._train_years[0] or np.NINF
-
-    def __train_max(self) -> Union[int, float]:
-        return self._train_years[1] or np.Inf
-
-    def __test_min(self) -> Union[int, float]:
-        return self._test_years[0] or np.NINF
-
-    def __test_max(self) -> Union[int, float]:
-        return self._test_years[1] or np.Inf
-
-    def __X(self, data_frame: pd.DataFrame) -> pd.DataFrame:
-        data_dummies = pd.get_dummies(self.data.select_dtypes("O"))
-        X_data = pd.get_dummies(data_frame.drop(["score", "oppo_score"], axis=1))
-
-        # Have to get missing dummy columns, because train & test years can have different
-        # teams/venues, resulting in data mismatch when trying to predict with a model
-        missing_cols = np.setdiff1d(data_dummies.columns, X_data.columns)
-        missing_df = pd.DataFrame(
-            {missing_col: 0 for missing_col in missing_cols}, index=X_data.index
-        )
-
-        return pd.concat([X_data, missing_df], axis=1).astype(float)
-
-    @staticmethod
-    def __y(data_frame: pd.DataFrame) -> pd.Series:
-        return data_frame["score"] - data_frame["oppo_score"]
-
-    @staticmethod
-    def __id_col(df):
-        return (
-            df["player_id"].astype(str)
-            + df["match_id"].astype(str)
-            + df["year"].astype(str)
-        )
+    def data(self) -> pd.DataFrame:
+        return self._data
 
     @staticmethod
     def __concat_data_frames(concated_data_frame, data_frame):
