+++ backend/server/ml_models/player_model/player_model.py
@@ -1,6 +1,6 @@
 """Model class trained on player data and its associated data class"""
 
-from typing import List, Tuple, Union, Sequence, Callable, Optional
+from typing import List, Sequence, Callable, Optional
 from functools import reduce
 import pandas as pd
 import numpy as np
@@ -8,7 +8,7 @@ from sklearn.preprocessing import StandardScaler
 from sklearn.base import BaseEstimator
 from xgboost import XGBRegressor
 
-from server.types import FeatureFunctionType, YearsType
+from server.types import FeatureFunctionType, YearPair
 from server.data_processors import (
     FeatureBuilder,
     PlayerDataStacker,
@@ -21,7 +21,7 @@ from server.data_processors.feature_functions import (
     add_cum_matches_played,
 )
 from server.data_processors import FitzroyDataReader
-from server.ml_models.ml_model import MLModel
+from server.ml_models.ml_model import MLModel, MLModelData
 
 MATCH_STATS_COLS = [
     "at_home",
@@ -32,12 +32,45 @@ MATCH_STATS_COLS = [
     "year",
     "round_number",
 ]
+DROPPABLE_COLS = [
+    "first_name",
+    "surname",
+    "round",
+    "local_start_time",
+    "attendance",
+    "hq1g",
+    "hq1b",
+    "hq2g",
+    "hq2b",
+    "hq3g",
+    "hq3b",
+    "hq4g",
+    "hq4b",
+    "aq1g",
+    "aq1b",
+    "aq2g",
+    "aq2b",
+    "aq3g",
+    "aq3b",
+    "aq4g",
+    "aq4b",
+    "jumper_no_",
+    "umpire_1",
+    "umpire_2",
+    "umpire_3",
+    "umpire_4",
+    "substitute",
+    "group_id",
+    "date",
+    "venue",
+]
 COL_TRANSLATIONS = {
     "season": "year",
     "time_on_ground__": "time_on_ground",
     "id": "player_id",
     "game": "match_id",
 }
+
 FEATURE_FUNCS: Sequence[FeatureFunctionType] = [
     add_last_year_brownlow_votes,
     add_rolling_player_stats,
@@ -72,36 +105,19 @@ class PlayerModel(MLModel):
         super().__init__(estimators=estimators, name=name, module_name=module_name)
 
 
-class PlayerModelData:
-    """Load and clean data for the XGBRegressor pipeline.
-
-    Args:
-        data_transformers (list[callable]): Functions that receive, transform,
-            and return data frames.
-        train_years (tuple[integer or None]): Minimum and maximum (inclusive) years
-            for the training data.
-        test_years (tuple[ingeter or None]): Minimum and maximum (inclusive) years
-            for the test data.
-
-    Attributes:
-        data (pandas.DataFrame): Cleaned, unfiltered data frame.
-        train_years (tuple[integer or None]): Minimum and maximum (inclusive) years
-            for the training data.
-        test_years (tuple[ingeter or None]): Minimum and maximum (inclusive) years
-            for the test data.
-    """
+class PlayerModelData(MLModelData):
+    """Load and clean player data"""
 
     def __init__(
         self,
         data_readers: List[Callable] = DATA_READERS,
         data_transformers: List[FeatureFunctionType] = DATA_TRANSFORMERS,
-        train_years: YearsType = (None, 2015),
-        test_years: YearsType = (2016, 2016),
+        train_years: YearPair = (None, 2015),
+        test_years: YearPair = (2016, 2016),
         start_date="1965-01-01",
         end_date="2016-12-31",
     ) -> None:
-        self._train_years = train_years
-        self._test_years = test_years
+        super().__init__(train_years=train_years, test_years=test_years)
 
         # Need to reverse the transformation steps, because composition makes the output
         # of each new function the argument for the previous
@@ -134,41 +150,7 @@ class PlayerModelData:
                 # player_id/match_id combos, decades apart, that by chance overlap
                 id=self.__id_col,
             )
-            .drop(
-                [
-                    "first_name",
-                    "surname",
-                    "round",
-                    "local_start_time",
-                    "attendance",
-                    "hq1g",
-                    "hq1b",
-                    "hq2g",
-                    "hq2b",
-                    "hq3g",
-                    "hq3b",
-                    "hq4g",
-                    "hq4b",
-                    "aq1g",
-                    "aq1b",
-                    "aq2g",
-                    "aq2b",
-                    "aq3g",
-                    "aq3b",
-                    "aq4g",
-                    "aq4b",
-                    "jumper_no_",
-                    "umpire_1",
-                    "umpire_2",
-                    "umpire_3",
-                    "umpire_4",
-                    "substitute",
-                    "group_id",
-                    "date",
-                    "venue",
-                ],
-                axis=1,
-            )
+            .drop(DROPPABLE_COLS, axis=1)
             # Some early matches (1800s) have fully-duplicated rows
             .drop_duplicates()
             .set_index("id")
@@ -192,95 +174,17 @@ class PlayerModelData:
             & (~data_frame["match_id"].isin(duplicate_matches))
         ]
 
-        self.data = compose_all(data_frame).dropna()
-
-    def train_data(self) -> Tuple[pd.DataFrame, pd.DataFrame]:
-        """Filter data by year to produce training data.
-
-        Returns:
-            Tuple[pandas.DataFrame]: Training features and labels.
-        """
-
-        data_train = self.data[
-            (self.data["year"] >= self.__train_min())
-            & (self.data["year"] <= self.__train_max())
-        ]
-
-        X_train = self.__X(data_train)
-        y_train = self.__y(data_train)
-
-        return X_train, y_train
-
-    def test_data(self) -> Tuple[pd.DataFrame, pd.DataFrame]:
-        """Filter data by year to produce test data.
-
-        Returns:
-            Tuple[pandas.DataFrame]: Test features and labels.
-        """
-
-        data_test = self.data[
-            (self.data["year"] >= self.__test_min())
-            & (self.data["year"] <= self.__test_max())
-        ]
-        X_test = self.__X(data_test)
-        y_test = self.__y(data_test)
-
-        return X_test, y_test
+        self._data = compose_all(data_frame).dropna()
 
     @property
-    def train_years(self) -> YearsType:
-        """Range of years for slicing training data"""
-
-        return self._train_years
-
-    @train_years.setter
-    def train_years(self, years: YearsType) -> None:
-        self._train_years = years
-
-    @property
-    def test_years(self) -> YearsType:
-        """Range of years for slicing test data"""
-
-        return self._test_years
-
-    @test_years.setter
-    def test_years(self, years: YearsType) -> None:
-        self._test_years = years
-
-    def __train_min(self) -> Union[int, float]:
-        return self._train_years[0] or np.NINF
-
-    def __train_max(self) -> Union[int, float]:
-        return self._train_years[1] or np.Inf
-
-    def __test_min(self) -> Union[int, float]:
-        return self._test_years[0] or np.NINF
-
-    def __test_max(self) -> Union[int, float]:
-        return self._test_years[1] or np.Inf
-
-    def __X(self, data_frame: pd.DataFrame) -> pd.DataFrame:
-        data_dummies = pd.get_dummies(self.data.select_dtypes("O"))
-        X_data = pd.get_dummies(data_frame.drop(["score", "oppo_score"], axis=1))
-
-        # Have to get missing dummy columns, because train & test years can have different
-        # teams/venues, resulting in data mismatch when trying to predict with a model
-        missing_cols = np.setdiff1d(data_dummies.columns, X_data.columns)
-        missing_df = pd.DataFrame(
-            {missing_col: 0 for missing_col in missing_cols}, index=X_data.index
-        )
-
-        return pd.concat([X_data, missing_df], axis=1).astype(float)
+    def data(self) -> pd.DataFrame:
+        return self._data
 
     @staticmethod
     def __compose_two(
         composed_func: FeatureFunctionType, func_element: FeatureFunctionType
     ) -> FeatureFunctionType:
         return lambda x: composed_func(func_element(x))
-
-    @staticmethod
-    def __y(data_frame: pd.DataFrame) -> pd.Series:
-        return data_frame["score"] - data_frame["oppo_score"]
 
     @staticmethod
     def __id_col(df):
